fix: further extension cleanup + theme fixes
